requirements.txt
django==2.0.3
djangorestframework==3.7.7
django-rest-swagger==2.1.2